+++ CestaDeCompras/CestaCompra.Data/Models/Consumidor.cs
@@ -12,7 +12,7 @@ namespace CestaCompra.Data.Models
         public int idconsumidor { get; set; }
         public int idpessoa { get; set; }
         public int datacadastro { get; set; }
-        public double admnistrador { get; set; }
+        public int admnistrador { get; set; }
         public string login { get; set; }
         public string senha { get; set; }
         public string nivel { get; set; }
